src/ refactor find_line_in_block find_sysline_in_block partial
For `find_line_in_block` when no `Line` is found, do return a
"partial" Line to the caller that is the from newline A to
end of the Block.
For `find_sysline_in_block` attempt to match a datetime on that
"partial" Line.
This is to allow matching long lines wihtin a blockzero analysis,
for long lines that extend into the next block (but those functions
will not look at). A "partial" means "looks very likely this
remaining data is a Line or Sysline".

Only used during blockzero analysis.

Issue #22