syslinereader.rs fix sort of indexes
The sorting of DateTimeParseDatasIndexes (`dt_patterns_indexes`)
was not actually occuring because ``dt_patterns_indexes` was
a `HashSet` which does not keep order (doh!).
Change the type to `Vec`. Assume no duplicate indexes will be
added to the `Vec` instance.

Add tests in syslinereader_tests.rs that use the
`DTPD!._test_cases` in `DATETIME_PARSE_DATAS` entries.