blockreader.rs attempt to parse more of the XZ header
Attempt to parse more of the XZ header and block #0 header.
Unfortunately, I couldn't figure get this working entirely. Leaving
the code in place as it does function.
The intent was to compensate for lzma-rs reading the entire file
during xz_decompress. However, that's a larger problem, see
https://github.com/gendx/lzma-rs/issues/110

Issue #12
Issue 283